Added platforms (WIP)
Added the first version of platforms. These platforms are made from 2 collision shapes, one at the top and one at the bottom.

The plan is, that when the bottom collision shape is hit, Bucket's Y-velocity will be inverted, but when Bucket hits the top collision shape, Bucket's Y-velocity will be zeroed.